More clean up and fix "should post to iframe to get bound on move contentlet and show bounds"
