+++ core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/edit-ema-editor.component.html
@@ -96,7 +96,7 @@
         (dragEnd)="onDragEnd($event)"
         data-testId="palette" />
 
-    @if( es.editorData.canEditVariant && es.editorData.mode === editorMode.EDIT) {
+    @if(es.editorData.canEditVariant && es.editorData.mode === editorMode.EDIT) {
     <dot-edit-ema-dialog
         #dialog
         (action)="onCustomEvent($event)"
